move tile decoding into backend api
