+++ src/lib/audio-render-tiles.ts
@@ -1,71 +1,6 @@
 import { BufferImageSource, Texture } from "pixi.js";
 
-const WAVEFORM_HEADER_BYTES = 24;
-const SPECTROGRAM_HEADER_BYTES = 40;
-
-export type WaveformTile = {
-  revision: number;
-  binCount: number;
-  samplesPerBin: number;
-  tileIndex: number;
-  min: Float32Array;
-  max: Float32Array;
-  representative: Float32Array;
-};
-
-export type SpectrogramTile = {
-  revision: number;
-  width: number;
-  height: number;
-  levelX: number;
-  levelY: number;
-  tileX: number;
-  tileY: number;
-  originX: number;
-  originY: number;
-  rgba: Uint8Array;
-};
-
-const asArrayBuffer = (value: ArrayBuffer | Uint8Array) => {
-  if (value instanceof ArrayBuffer) return value;
-  return value.buffer.slice(value.byteOffset, value.byteOffset + value.byteLength) as ArrayBuffer;
-};
-
-export function decodeWaveformTile(value: ArrayBuffer | Uint8Array): WaveformTile {
-  const buffer = asArrayBuffer(value);
-  const view = new DataView(buffer);
-  const revision = Number(view.getBigUint64(0, true));
-  const binCount = view.getUint32(8, true);
-  const samplesPerBin = view.getUint32(12, true);
-  const tileIndex = view.getUint32(16, true);
-  const min = new Float32Array(binCount);
-  const max = new Float32Array(binCount);
-  const representative = new Float32Array(binCount);
-  for (let i = 0; i < binCount; i += 1) {
-    const offset = WAVEFORM_HEADER_BYTES + i * 12;
-    min[i] = view.getFloat32(offset, true);
-    max[i] = view.getFloat32(offset + 4, true);
-    representative[i] = view.getFloat32(offset + 8, true);
-  }
-  return { revision, binCount, samplesPerBin, tileIndex, min, max, representative };
-}
-
-export function decodeSpectrogramTile(value: ArrayBuffer | Uint8Array): SpectrogramTile {
-  const buffer = asArrayBuffer(value);
-  const view = new DataView(buffer);
-  return {
-    revision: Number(view.getBigUint64(0, true)),
-    width: view.getUint32(8, true),
-    height: view.getUint32(12, true),
-    levelX: view.getUint32(16, true),
-    levelY: view.getUint32(20, true),
-    tileX: view.getUint32(24, true),
-    tileY: view.getUint32(28, true),
-    originX: view.getUint32(32, true),
-    originY: view.getUint32(36, true),
-    rgba: new Uint8Array(buffer, SPECTROGRAM_HEADER_BYTES),
-  };
-}
+import type { SpectrogramTile, WaveformTile } from "../api";
 
 type TextureCacheEntry = {
   texture: Texture;
